longtitude and latitude become float field

+++ location/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 # Create your models here.
 class BinLocation(models.Model):
-    loc_lng = models.DecimalField(max_digits = 5, decimal_places = 2)
-    loc_lat = models.DecimalField(max_digits = 5, decimal_places = 2)
+    loc_lng = models.FloatField()
+    loc_lat = models.FloatField()
     building = models.TextField()
     room = models.IntegerField()
